refactor: update logger import paths across multiple services and consolidate model_config in SettingsService
