+++ core/services/core/bootstrapper_service.py
@@ -4,7 +4,7 @@ from core.routers import routes
 from flow_engine.flow_chain.services.flow_node_registry_service import FlowNodeRegistry
 from shared.services.database import database_service
 from shared.services.msg_broker import MessageBrokerService
-from shared.utils import logger
+from shared.utils.logger import logger
 
 
 class BootstrapperService:
